[v0.1.0] Major Refactor
- Remove moment and Snap dependencies
- Use Rollup as build tool
- Use Prettier for linting/styling
- Use Jest for testing
- Use yarn
