+++ .babelrc
@@ -1,4 +1,3 @@
 {
-  "presets": ["es2015"],
-  "plugins": ["babel-plugin-add-module-exports"]
+  "presets": ["env"]
 }
