+++ src/Arrow.js
@@ -1,105 +1,96 @@
-/* global Snap */
-/*
-	Class: Arrow
-	from_task ---> to_task
+import { createSVG } from './svg_utils';
 
-	Opts:
-		gantt (Gantt object)
-		from_task (Bar object)
-		to_task (Bar object)
-*/
+export default class Arrow {
+    constructor(gantt, from_task, to_task) {
+        this.gantt = gantt;
+        this.from_task = from_task;
+        this.to_task = to_task;
 
-export default function Arrow(gt, from_task, to_task) {
+        this.calculate_path();
+        this.draw();
+    }
 
-	const self = {};
+    calculate_path() {
+        let start_x =
+            this.from_task.$bar.getX() + this.from_task.$bar.getWidth() / 2;
 
-	function init() {
-		self.from_task = from_task;
-		self.to_task = to_task;
-		prepare();
-		draw();
-	}
+        const condition = () =>
+            this.to_task.$bar.getX() < start_x + this.gantt.options.padding &&
+            start_x > this.from_task.$bar.getX() + this.gantt.options.padding;
 
-	function prepare() {
+        while (condition()) {
+            start_x -= 10;
+        }
 
-		self.start_x = from_task.$bar.getX() + from_task.$bar.getWidth() / 2;
+        const start_y =
+            this.gantt.options.header_height +
+            this.gantt.options.bar.height +
+            (this.gantt.options.padding + this.gantt.options.bar.height) *
+                this.from_task.task._index +
+            this.gantt.options.padding;
 
-		const condition = () =>
-			to_task.$bar.getX() < self.start_x + gt.config.padding &&
-				self.start_x > from_task.$bar.getX() + gt.config.padding;
+        const end_x = this.to_task.$bar.getX() - this.gantt.options.padding / 2;
+        const end_y =
+            this.gantt.options.header_height +
+            this.gantt.options.bar.height / 2 +
+            (this.gantt.options.padding + this.gantt.options.bar.height) *
+                this.to_task.task._index +
+            this.gantt.options.padding;
 
-		while(condition()) {
-			self.start_x -= 10;
-		}
+        const from_is_below_to =
+            this.from_task.task._index > this.to_task.task._index;
+        const curve = this.gantt.options.arrow.curve;
+        const clockwise = from_is_below_to ? 1 : 0;
+        const curve_y = from_is_below_to ? -curve : curve;
+        const offset = from_is_below_to
+            ? end_y + this.gantt.options.arrow.curve
+            : end_y - this.gantt.options.arrow.curve;
 
-		self.start_y = gt.config.header_height + gt.config.bar.height +
-			(gt.config.padding + gt.config.bar.height) * from_task.task._index +
-			gt.config.padding;
+        this.path = `
+            M ${start_x} ${start_y}
+            V ${offset}
+            a ${curve} ${curve} 0 0 ${clockwise} ${curve} ${curve_y}
+            L ${end_x} ${end_y}
+            m -5 -5
+            l 5 5
+            l -5 5`;
 
-		self.end_x = to_task.$bar.getX() - gt.config.padding / 2;
-		self.end_y = gt.config.header_height + gt.config.bar.height / 2 +
-			(gt.config.padding + gt.config.bar.height) * to_task.task._index +
-			gt.config.padding;
+        if (
+            this.to_task.$bar.getX() <
+            this.from_task.$bar.getX() + this.gantt.options.padding
+        ) {
+            const down_1 = this.gantt.options.padding / 2 - curve;
+            const down_2 =
+                this.to_task.$bar.getY() +
+                this.to_task.$bar.getHeight() / 2 -
+                curve_y;
+            const left = this.to_task.$bar.getX() - this.gantt.options.padding;
 
-		const from_is_below_to = (from_task.task._index > to_task.task._index);
-		self.curve = gt.config.arrow.curve;
-		self.clockwise = from_is_below_to ? 1 : 0;
-		self.curve_y = from_is_below_to ? -self.curve : self.curve;
-		self.offset = from_is_below_to ?
-			self.end_y + gt.config.arrow.curve :
-			self.end_y - gt.config.arrow.curve;
+            this.path = `
+                M ${start_x} ${start_y}
+                v ${down_1}
+                a ${curve} ${curve} 0 0 1 -${curve} ${curve}
+                H ${left}
+                a ${curve} ${curve} 0 0 ${clockwise} -${curve} ${curve_y}
+                V ${down_2}
+                a ${curve} ${curve} 0 0 ${clockwise} ${curve} ${curve_y}
+                L ${end_x} ${end_y}
+                m -5 -5
+                l 5 5
+                l -5 5`;
+        }
+    }
 
-		self.path =
-			Snap.format('M {start_x} {start_y} V {offset} ' +
-				'a {curve} {curve} 0 0 {clockwise} {curve} {curve_y} ' +
-				'L {end_x} {end_y} m -5 -5 l 5 5 l -5 5',
-				{
-					start_x: self.start_x,
-					start_y: self.start_y,
-					end_x: self.end_x,
-					end_y: self.end_y,
-					offset: self.offset,
-					curve: self.curve,
-					clockwise: self.clockwise,
-					curve_y: self.curve_y
-				});
+    draw() {
+        this.element = createSVG('path', {
+            d: this.path,
+            'data-from': this.from_task.task.id,
+            'data-to': this.to_task.task.id
+        });
+    }
 
-		if(to_task.$bar.getX() < from_task.$bar.getX() + gt.config.padding) {
-			self.path =
-				Snap.format('M {start_x} {start_y} v {down_1} ' +
-				'a {curve} {curve} 0 0 1 -{curve} {curve} H {left} ' +
-				'a {curve} {curve} 0 0 {clockwise} -{curve} {curve_y} V {down_2} ' +
-				'a {curve} {curve} 0 0 {clockwise} {curve} {curve_y} ' +
-				'L {end_x} {end_y} m -5 -5 l 5 5 l -5 5',
-					{
-						start_x: self.start_x,
-						start_y: self.start_y,
-						end_x: self.end_x,
-						end_y: self.end_y,
-						down_1: gt.config.padding / 2 - self.curve,
-						down_2: to_task.$bar.getY() + to_task.$bar.getHeight() / 2 - self.curve_y,
-						left: to_task.$bar.getX() - gt.config.padding,
-						offset: self.offset,
-						curve: self.curve,
-						clockwise: self.clockwise,
-						curve_y: self.curve_y
-					});
-		}
-	}
-
-	function draw() {
-		self.element = gt.canvas.path(self.path)
-			.attr('data-from', self.from_task.task.id)
-			.attr('data-to', self.to_task.task.id);
-	}
-
-	function update() { // eslint-disable-line
-		prepare();
-		self.element.attr('d', self.path);
-	}
-	self.update = update;
-
-	init();
-
-	return self;
+    update() {
+        this.calculate_path();
+        this.element.setAttribute('d', this.path);
+    }
 }
