+++ src/index.js
@@ -1,867 +1,12 @@
-var Gantt = (function () {
-'use strict';
-
-const YEAR = 'year';
-const MONTH = 'month';
-const DAY = 'day';
-const HOUR = 'hour';
-const MINUTE = 'minute';
-const SECOND = 'second';
-const MILLISECOND = 'millisecond';
-
-const month_names = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December'
-];
-
-var date_utils = {
-    parse(date, date_separator = '-', time_separator = ':') {
-        if (date instanceof Date) {
-            return date;
-        }
-        if (typeof date === 'string') {
-            let date_parts, time_parts;
-            const parts = date.split(' ');
-
-            date_parts = parts[0]
-                .split(date_separator)
-                .map(val => parseInt(val, 10));
-            time_parts = parts[1] && parts[1].split(time_separator);
-
-            // month is 0 indexed
-            date_parts[1] = date_parts[1] - 1;
-
-            let vals = date_parts;
-
-            if (time_parts && time_parts.length) {
-                vals = vals.concat(time_parts);
-            }
-
-            return new Date(...vals);
-        }
-    },
-
-    to_string(date, with_time = false) {
-        if (!(date instanceof Date)) {
-            throw new TypeError('Invalid argument type');
-        }
-        const vals = this.get_date_values(date).map((val, i) => {
-            if (i === 1) {
-                // add 1 for month
-                val = val + 1;
-            }
-
-            return padStart(val + '', 2, '0');
-        });
-        const date_string = `${vals[0]}-${vals[1]}-${vals[2]}`;
-        const time_string = `${vals[3]}:${vals[4]}:${vals[5]}`;
-
-        return date_string + (with_time ? ' ' + time_string : '');
-    },
-
-    format(date, format_string = 'YYYY-MM-DD HH:mm:ss') {
-        const values = this.get_date_values(date).map(d => padStart(d, 2, 0));
-        const format_map = {
-            YYYY: values[0],
-            MM: padStart(+values[1] + 1, 2, 0),
-            DD: values[2],
-            HH: values[3],
-            mm: values[4],
-            ss: values[5],
-            D: values[2],
-            MMMM: month_names[+values[1]],
-            MMM: month_names[+values[1]]
-        };
-
-        let str = format_string;
-
-        Object.keys(format_map)
-            .sort((a, b) => b.length - a.length) // big string first
-            .forEach(key => {
-                str = str.replace(key, format_map[key]);
-            });
-
-        return str;
-    },
-
-    diff(date_a, date_b, scale = DAY) {
-        let milliseconds, seconds, hours, minutes, days, months, years;
-
-        milliseconds = date_a - date_b;
-        seconds = milliseconds / 1000;
-        minutes = seconds / 60;
-        hours = minutes / 60;
-        days = hours / 24;
-        months = days / 30;
-        years = months / 12;
-
-        if (!scale.endsWith('s')) {
-            scale += 's';
-        }
-
-        return Math.floor(
-            {
-                milliseconds,
-                seconds,
-                minutes,
-                hours,
-                days,
-                months,
-                years
-            }[scale]
-        );
-    },
-
-    today() {
-        const vals = this.get_date_values(new Date()).slice(0, 3);
-        return new Date(...vals);
-    },
-
-    now() {
-        return new Date();
-    },
-
-    add(date, qty, scale) {
-        qty = parseInt(qty, 10);
-        const vals = [
-            date.getFullYear() + (scale === YEAR ? qty : 0),
-            date.getMonth() + (scale === MONTH ? qty : 0),
-            date.getDate() + (scale === DAY ? qty : 0),
-            date.getHours() + (scale === HOUR ? qty : 0),
-            date.getMinutes() + (scale === MINUTE ? qty : 0),
-            date.getSeconds() + (scale === SECOND ? qty : 0),
-            date.getMilliseconds() + (scale === MILLISECOND ? qty : 0)
-        ];
-        return new Date(...vals);
-    },
-
-    start_of(date, scale) {
-        const scores = {
-            [YEAR]: 6,
-            [MONTH]: 5,
-            [DAY]: 4,
-            [HOUR]: 3,
-            [MINUTE]: 2,
-            [SECOND]: 1,
-            [MILLISECOND]: 0
-        };
-
-        function should_reset(_scale) {
-            const max_score = scores[scale];
-            return scores[_scale] <= max_score;
-        }
-
-        const vals = [
-            date.getFullYear(),
-            should_reset(YEAR) ? 0 : date.getMonth(),
-            should_reset(MONTH) ? 1 : date.getDate(),
-            should_reset(DAY) ? 0 : date.getHours(),
-            should_reset(HOUR) ? 0 : date.getMinutes(),
-            should_reset(MINUTE) ? 0 : date.getSeconds(),
-            should_reset(SECOND) ? 0 : date.getMilliseconds()
-        ];
-
-        return new Date(...vals);
-    },
-
-    clone(date) {
-        return new Date(...this.get_date_values(date));
-    },
-
-    get_date_values(date) {
-        return [
-            date.getFullYear(),
-            date.getMonth(),
-            date.getDate(),
-            date.getHours(),
-            date.getMinutes(),
-            date.getSeconds(),
-            date.getMilliseconds()
-        ];
-    },
-
-    get_days_in_month(date) {
-        const no_of_days = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
-        const month = date.getMonth();
-
-        if (month !== 1) {
-            return no_of_days[month];
-        }
-
-        // Feb
-        const year = date.getFullYear();
-        if ((year % 4 == 0 && year % 100 != 0) || year % 400 == 0) {
-            return 29;
-        }
-        return 28;
-    }
-};
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/padStart
-function padStart(str, targetLength, padString) {
-    str = str + '';
-    targetLength = targetLength >> 0;
-    padString = String(typeof padString !== 'undefined' ? padString : ' ');
-    if (str.length > targetLength) {
-        return String(str);
-    } else {
-        targetLength = targetLength - str.length;
-        if (targetLength > padString.length) {
-            padString += padString.repeat(targetLength / padString.length);
-        }
-        return padString.slice(0, targetLength) + String(str);
-    }
-}
-
-function $(expr, con) {
-    return typeof expr === 'string'
-        ? (con || document).querySelector(expr)
-        : expr || null;
-}
-
-function createSVG(tag, attrs) {
-    const elem = document.createElementNS('http://www.w3.org/2000/svg', tag);
-    for (let attr in attrs) {
-        if (attr === 'append_to') {
-            const parent = attrs.append_to;
-            parent.appendChild(elem);
-        } else if (attr === 'innerHTML') {
-            elem.innerHTML = attrs.innerHTML;
-        } else {
-            elem.setAttribute(attr, attrs[attr]);
-        }
-    }
-    return elem;
-}
-
-
-
-$.on = (element, event, selector, callback) => {
-    if (!callback) {
-        callback = selector;
-        $.bind(element, event, callback);
-    } else {
-        $.delegate(element, event, selector, callback);
-    }
-};
-
-$.off = (element, event, handler) => {
-    element.removeEventListener(event, handler);
-};
-
-$.bind = (element, event, callback) => {
-    event.split(/\s+/).forEach(function(event) {
-        element.addEventListener(event, callback);
-    });
-};
-
-$.delegate = (element, event, selector, callback) => {
-    element.addEventListener(event, function(e) {
-        const delegatedTarget = e.target.closest(selector);
-        if (delegatedTarget) {
-            e.delegatedTarget = delegatedTarget;
-            callback.call(this, e, delegatedTarget);
-        }
-    });
-};
-
-$.closest = (selector, element) => {
-    if (!element) return null;
-
-    if (element.matches(selector)) {
-        return element;
-    }
-
-    return $.closest(selector, element.parentNode);
-};
-
-$.attr = (element, attr, value) => {
-    if (!value && typeof attr === 'string') {
-        return element.getAttribute(attr);
-    }
-
-    if (typeof attr === 'object') {
-        for (let key in attr) {
-            $.attr(element, key, attr[key]);
-        }
-        return;
-    }
-
-    element.setAttribute(attr, value);
-};
-
-class Bar {
-    constructor(gantt, task) {
-        this.set_defaults(gantt, task);
-        this.prepare();
-        this.draw();
-        this.bind();
-    }
-
-    set_defaults(gantt, task) {
-        this.action_completed = false;
-        this.gantt = gantt;
-        this.task = task;
-    }
-
-    prepare() {
-        this.prepare_values();
-        this.prepare_helpers();
-    }
-
-    prepare_values() {
-        this.invalid = this.task.invalid;
-        this.height = this.gantt.options.bar.height;
-        this.x = this.compute_x();
-        this.y = this.compute_y();
-        this.corner_radius = 3;
-        this.duration =
-            (date_utils.diff(this.task._end, this.task._start, 'hour') + 24) /
-            this.gantt.options.step;
-        this.width = this.gantt.options.column_width * this.duration;
-        this.progress_width =
-            this.gantt.options.column_width *
-                this.duration *
-                (this.task.progress / 100) || 0;
-        this.group = createSVG('g', {
-            class: 'bar-wrapper ' + (this.task.custom_class || ''),
-            'data-id': this.task.id
-        });
-        this.bar_group = createSVG('g', {
-            class: 'bar-group',
-            append_to: this.group
-        });
-        this.handle_group = createSVG('g', {
-            class: 'handle-group',
-            append_to: this.group
-        });
-    }
-
-    prepare_helpers() {
-        SVGElement.prototype.getX = function() {
-            return +this.getAttribute('x');
-        };
-        SVGElement.prototype.getY = function() {
-            return +this.getAttribute('y');
-        };
-        SVGElement.prototype.getWidth = function() {
-            return +this.getAttribute('width');
-        };
-        SVGElement.prototype.getHeight = function() {
-            return +this.getAttribute('height');
-        };
-        SVGElement.prototype.getEndX = function() {
-            return this.getX() + this.getWidth();
-        };
-    }
-
-    draw() {
-        this.draw_bar();
-        this.draw_progress_bar();
-        this.draw_label();
-        this.draw_resize_handles();
-    }
-
-    draw_bar() {
-        this.$bar = createSVG('rect', {
-            x: this.x,
-            y: this.y,
-            width: this.width,
-            height: this.height,
-            rx: this.corner_radius,
-            ry: this.corner_radius,
-            class: 'bar',
-            append_to: this.bar_group
-        });
-
-        if (this.invalid) {
-            this.$bar.classList.add('bar-invalid');
-        }
-    }
-
-    draw_progress_bar() {
-        if (this.invalid) return;
-        this.$bar_progress = createSVG('rect', {
-            x: this.x,
-            y: this.y,
-            width: this.progress_width,
-            height: this.height,
-            rx: this.corner_radius,
-            ry: this.corner_radius,
-            class: 'bar-progress',
-            append_to: this.bar_group
-        });
-    }
-
-    draw_label() {
-        createSVG('text', {
-            x: this.x + this.width / 2,
-            y: this.y + this.height / 2,
-            innerHTML: this.task.name,
-            class: 'bar-label',
-            append_to: this.bar_group
-        });
-        this.update_label_position();
-    }
-
-    draw_resize_handles() {
-        if (this.invalid) return;
-
-        const bar = this.$bar;
-        const handle_width = 8;
-
-        createSVG('rect', {
-            x: bar.getX() + bar.getWidth() - 9,
-            y: bar.getY() + 1,
-            width: handle_width,
-            height: this.height - 2,
-            rx: this.corner_radius,
-            ry: this.corner_radius,
-            class: 'handle right',
-            append_to: this.handle_group
-        });
-
-        createSVG('rect', {
-            x: bar.getX() + 1,
-            y: bar.getY() + 1,
-            width: handle_width,
-            height: this.height - 2,
-            rx: this.corner_radius,
-            ry: this.corner_radius,
-            class: 'handle left',
-            append_to: this.handle_group
-        });
-
-        if (this.task.progress && this.task.progress < 100) {
-            this.$handle_progress = createSVG('polygon', {
-                points: this.get_progress_polygon_points().join(','),
-                class: 'handle progress',
-                append_to: this.handle_group
-            });
-        }
-    }
-
-    get_progress_polygon_points() {
-        const bar_progress = this.$bar_progress;
-        return [
-            bar_progress.getEndX() - 5,
-            bar_progress.getY() + bar_progress.getHeight(),
-            bar_progress.getEndX() + 5,
-            bar_progress.getY() + bar_progress.getHeight(),
-            bar_progress.getEndX(),
-            bar_progress.getY() + bar_progress.getHeight() - 8.66
-        ];
-    }
-
-    bind() {
-        if (this.invalid) return;
-        this.setup_click_event();
-        this.show_details();
-        // this.bind_resize_progress();
-    }
-
-    show_details() {
-        this.group.onclick = e => {
-            if (this.action_completed) {
-                // just finished a move action, wait for a few seconds
-                return;
-            }
-
-            const start_date = date_utils.format(this.task._start, 'MMM D');
-            const end_date = date_utils.format(this.task._end, 'MMM D');
-            const subtitle = start_date + ' - ' + end_date;
-
-            this.gantt.show_popup({
-                target_element: this.$bar,
-                title: this.task.name,
-                subtitle: subtitle
-            });
-        };
-    }
-
-    update_bar_position({ x = null, width = null }) {
-        const bar = this.$bar;
-        if (x) {
-            // get all x values of parent task
-            const xs = this.task.dependencies.map(dep => {
-                return this.gantt.get_bar(dep).$bar.getX();
-            });
-            // child task must not go before parent
-            const valid_x = xs.reduce((prev, curr) => {
-                return x >= curr;
-            }, x);
-            if (!valid_x) {
-                width = null;
-                return;
-            }
-            this.update_attr(bar, 'x', x);
-        }
-        if (width && width >= this.gantt.options.column_width) {
-            this.update_attr(bar, 'width', width);
-        }
-        this.update_label_position();
-        this.update_handle_position();
-        this.update_progressbar_position();
-        this.update_arrow_position();
-        // this.update_details_position();
-    }
-
-    setup_click_event() {
-        this.group.onclick = () => {
-            if (this.action_completed) {
-                // just finished a move action, wait for a few seconds
-                return;
-            }
-            if (this.group.classList.contains('active')) {
-                this.gantt.trigger_event('click', [this.task]);
-            }
-            this.gantt.unselect_all();
-            this.group.classList.toggle('active');
-        };
-    }
-
-    date_changed() {
-        const { new_start_date, new_end_date } = this.compute_start_end_date();
-        this.task._start = new_start_date;
-        this.task._end = new_end_date;
-
-        this.gantt.trigger_event('date_change', [
-            this.task,
-            new_start_date,
-            new_end_date
-        ]);
-    }
-
-    progress_changed() {
-        const new_progress = this.compute_progress();
-        this.task.progress = new_progress;
-        this.gantt.trigger_event('progress_change', [this.task, new_progress]);
-    }
-
-    set_action_completed() {
-        this.action_completed = true;
-        setTimeout(() => (this.action_completed = false), 2000);
-    }
-
-    compute_start_end_date() {
-        const bar = this.$bar;
-        const x_in_units = bar.getX() / this.gantt.options.column_width;
-        const new_start_date = date_utils.add(
-            this.gantt.gantt_start,
-            x_in_units * this.gantt.options.step,
-            'hours'
-        );
-        const width_in_units = bar.getWidth() / this.gantt.options.column_width;
-        const new_end_date = date_utils.add(
-            new_start_date,
-            width_in_units * this.gantt.options.step,
-            'hours'
-        );
-        // lets say duration is 2 days
-        // start_date = May 24 00:00:00
-        // end_date = May 24 + 2 days = May 26 (incorrect)
-        // so subtract 1 second so that
-        // end_date = May 25 23:59:59
-        date_utils.add(new_end_date, -1, 'second');
-        return { new_start_date, new_end_date };
-    }
-
-    compute_progress() {
-        const progress =
-            this.$bar_progress.getWidth() / this.$bar.getWidth() * 100;
-        return parseInt(progress, 10);
-    }
-
-    compute_x() {
-        let x =
-            date_utils.diff(this.task._start, this.gantt.gantt_start, 'hour') /
-            this.gantt.options.step *
-            this.gantt.options.column_width;
-
-        if (this.gantt.view_is('Month')) {
-            x =
-                date_utils.diff(
-                    this.task._start,
-                    this.gantt.gantt_start,
-                    'day'
-                ) *
-                this.gantt.options.column_width /
-                30;
-        }
-        return x;
-    }
-
-    compute_y() {
-        return (
-            this.gantt.options.header_height +
-            this.gantt.options.padding +
-            this.task._index * (this.height + this.gantt.options.padding)
-        );
-    }
-
-    get_snap_position(dx) {
-        let odx = dx,
-            rem,
-            position;
-
-        if (this.gantt.view_is('Week')) {
-            rem = dx % (this.gantt.options.column_width / 7);
-            position =
-                odx -
-                rem +
-                (rem < this.gantt.options.column_width / 14
-                    ? 0
-                    : this.gantt.options.column_width / 7);
-        } else if (this.gantt.view_is('Month')) {
-            rem = dx % (this.gantt.options.column_width / 30);
-            position =
-                odx -
-                rem +
-                (rem < this.gantt.options.column_width / 60
-                    ? 0
-                    : this.gantt.options.column_width / 30);
-        } else {
-            rem = dx % this.gantt.options.column_width;
-            position =
-                odx -
-                rem +
-                (rem < this.gantt.options.column_width / 2
-                    ? 0
-                    : this.gantt.options.column_width);
-        }
-        return position;
-    }
-
-    update_attr(element, attr, value) {
-        value = +value;
-        if (!isNaN(value)) {
-            element.setAttribute(attr, value);
-        }
-        return element;
-    }
-
-    update_progressbar_position() {
-        this.$bar_progress.setAttribute('x', this.$bar.getX());
-        this.$bar_progress.setAttribute(
-            'width',
-            this.$bar.getWidth() * (this.task.progress / 100)
-        );
-    }
-
-    update_label_position() {
-        const bar = this.$bar,
-            label = this.group.querySelector('.bar-label');
-        if (label.getBBox().width > bar.getWidth()) {
-            label.classList.add('big');
-            label.setAttribute('x', bar.getX() + bar.getWidth() + 5);
-        } else {
-            label.classList.remove('big');
-            label.setAttribute('x', bar.getX() + bar.getWidth() / 2);
-        }
-    }
-
-    update_handle_position() {
-        const bar = this.$bar;
-        this.handle_group
-            .querySelector('.handle.left')
-            .setAttribute('x', bar.getX() + 1);
-        this.handle_group
-            .querySelector('.handle.right')
-            .setAttribute('x', bar.getEndX() - 9);
-        const handle = this.group.querySelector('.handle.progress');
-        handle &&
-            handle.setAttribute('points', this.get_progress_polygon_points());
-    }
-
-    update_arrow_position() {
-        this.arrows = this.arrows || [];
-        for (let arrow of this.arrows) {
-            arrow.update();
-        }
-    }
-
-    update_details_position() {
-        const { x, y } = get_details_position();
-        this.details_box && this.details_box.transform(`t${x},${y}`);
-    }
-}
-
-class Arrow {
-    constructor(gantt, from_task, to_task) {
-        this.gantt = gantt;
-        this.from_task = from_task;
-        this.to_task = to_task;
-
-        this.calculate_path();
-        this.draw();
-    }
-
-    calculate_path() {
-        let start_x =
-            this.from_task.$bar.getX() + this.from_task.$bar.getWidth() / 2;
-
-        const condition = () =>
-            this.to_task.$bar.getX() < start_x + this.gantt.options.padding &&
-            start_x > this.from_task.$bar.getX() + this.gantt.options.padding;
-
-        while (condition()) {
-            start_x -= 10;
-        }
-
-        const start_y =
-            this.gantt.options.header_height +
-            this.gantt.options.bar.height +
-            (this.gantt.options.padding + this.gantt.options.bar.height) *
-                this.from_task.task._index +
-            this.gantt.options.padding;
-
-        const end_x = this.to_task.$bar.getX() - this.gantt.options.padding / 2;
-        const end_y =
-            this.gantt.options.header_height +
-            this.gantt.options.bar.height / 2 +
-            (this.gantt.options.padding + this.gantt.options.bar.height) *
-                this.to_task.task._index +
-            this.gantt.options.padding;
-
-        const from_is_below_to =
-            this.from_task.task._index > this.to_task.task._index;
-        const curve = this.gantt.options.arrow.curve;
-        const clockwise = from_is_below_to ? 1 : 0;
-        const curve_y = from_is_below_to ? -curve : curve;
-        const offset = from_is_below_to
-            ? end_y + this.gantt.options.arrow.curve
-            : end_y - this.gantt.options.arrow.curve;
-
-        this.path = `
-            M ${start_x} ${start_y}
-            V ${offset}
-            a ${curve} ${curve} 0 0 ${clockwise} ${curve} ${curve_y}
-            L ${end_x} ${end_y}
-            m -5 -5
-            l 5 5
-            l -5 5`;
-
-        if (
-            this.to_task.$bar.getX() <
-            this.from_task.$bar.getX() + this.gantt.options.padding
-        ) {
-            const down_1 = this.gantt.options.padding / 2 - curve;
-            const down_2 =
-                this.to_task.$bar.getY() +
-                this.to_task.$bar.getHeight() / 2 -
-                curve_y;
-            const left = this.to_task.$bar.getX() - this.gantt.options.padding;
-
-            this.path = `
-                M ${start_x} ${start_y}
-                v ${down_1}
-                a ${curve} ${curve} 0 0 1 -${curve} ${curve}
-                H ${left}
-                a ${curve} ${curve} 0 0 ${clockwise} -${curve} ${curve_y}
-                V ${down_2}
-                a ${curve} ${curve} 0 0 ${clockwise} ${curve} ${curve_y}
-                L ${end_x} ${end_y}
-                m -5 -5
-                l 5 5
-                l -5 5`;
-        }
-    }
-
-    draw() {
-        this.element = createSVG('path', {
-            d: this.path,
-            'data-from': this.from_task.task.id,
-            'data-to': this.to_task.task.id
-        });
-    }
-
-    update() {
-        this.calculate_path();
-        this.element.setAttribute('d', this.path);
-    }
-}
-
-class Popup {
-    constructor(parent) {
-        this.parent = parent;
-        this.make();
-    }
-
-    make() {
-        this.parent.innerHTML = `
-            <div class="title"></div>
-            <div class="subtitle"></div>
-            <div class="pointer"></div>
-        `;
-
-        this.hide();
-
-        this.title = this.parent.querySelector('.title');
-        this.subtitle = this.parent.querySelector('.subtitle');
-        this.pointer = this.parent.querySelector('.pointer');
-    }
-
-    show(options) {
-        if (!options.target_element) {
-            throw new Error('target_element is required to show popup');
-        }
-        if (!options.position) {
-            options.position = 'left';
-        }
-        const target_element = options.target_element;
-
-        // set data
-        this.title.innerHTML = options.title;
-        this.subtitle.innerHTML = options.subtitle;
-
-        this.parent.style.width = this.parent.clientWidth + 'px';
-
-        // set position
-        let position_meta;
-        if (target_element instanceof HTMLElement) {
-            position_meta = target_element.getBoundingClientRect();
-        } else if (target_element instanceof SVGElement) {
-            position_meta = options.target_element.getBBox();
-        }
-
-        if (options.position === 'left') {
-            this.parent.style.left =
-                position_meta.x + (position_meta.width + 10) + 'px';
-            this.parent.style.top =
-                position_meta.y -
-                this.title.clientHeight / 2 +
-                position_meta.height / 2 +
-                'px';
-
-            this.pointer.style.transform = 'rotateZ(90deg)';
-            this.pointer.style.left = '-7px';
-            this.pointer.style.top =
-                this.title.clientHeight / 2 -
-                this.pointer.getBoundingClientRect().height +
-                'px';
-        }
-
-        // show
-        this.parent.style.opacity = 1;
-    }
-
-    hide() {
-        this.parent.style.opacity = 0;
-    }
-}
-
-class Gantt {
+import date_utils from './date_utils';
+import { $, createSVG } from './svg_utils';
+import Bar from './bar';
+import Arrow from './arrow';
+import Popup from './popup';
+
+import './gantt.scss';
+
+export default class Gantt {
     constructor(wrapper, tasks, options) {
         this.setup_wrapper(wrapper);
         this.setup_options(options);
@@ -1701,7 +846,3 @@ function generate_id(task) {
             .slice(2, 12)
     );
 }
-
-return Gantt;
-
-}());
